refactor(tests): split AnalysisStoreTests (926→≤232) (#272, #266)
Split MoolahTests/Features/AnalysisStoreTests.swift along @Suite
boundaries into six sibling files — each drops well under the
file_length 400-line threshold, eliminating the baseline entry.

While rewiring the suites, replace the baselined force-unwraps that
would otherwise re-fire in the new files:
  - Decimal(string:)! → Decimal(Int) (all values were integer-valued)
  - UserDefaults(suiteName:)! → try #require(…) via throwing helper
  - store.lastLoadedAt!, forecasts.last!, calendar.date(…)! → #require

Also rewrites the baselined multiline_parameters violation on the
balance() helper to one-parameter-per-line.

Stacked on #350 until it lands; retarget to main then.
